requirements.txt
backports.csv==1.0.7
defusedxml==0.6.0
diff-match-patch==20181111
dj-static==0.0.6
Django==2.2.5
django-ckeditor==5.7.1
django-import-export==1.2.0
django-js-asset==1.2.2
django-money==0.15
django-rest-framework==0.1.0
django-userforeignkey==0.3.0
djangorestframework==3.10.2
et-xmlfile==1.0.1
gunicorn==19.9.0
html5lib==1.0.1
jdcal==1.4.1
odfpy==1.4.0
openpyxl==2.6.3
Pillow==6.1.0
psycopg2==2.8.3
py-moneyed==0.8.0
PyPDF2==1.26.0
pytz==2019.2
PyYAML==5.1.2
reportlab==3.5.23
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
tablib==0.13.0
webencodings==0.5.1
xhtml2pdf==0.2.3
xlrd==1.2.0
xlwt==1.3.0
